requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==0.3.0
datadog==0.50.0
tenacity==9.0.0
python-dotenv==1.0.1
google-cloud-aiplatform
ddtrace
